+++ src/earf/rules/__init__.py
@@ -1,5 +1,12 @@
 from .models import RuleDefinition
-from .loader import RuleLoader
+from .loader import RuleLoader, YamlRuleLoader
+from .catalog import RuleCatalog
 from .engine import RuleEngine
 
-__all__ = ["RuleDefinition", "RuleLoader", "RuleEngine"]
+__all__ = [
+	"RuleDefinition",
+	"RuleLoader",
+	"YamlRuleLoader",
+	"RuleCatalog",
+	"RuleEngine",
+]